Fix computation of buy amount estimation (#14)
The semantics of the API changed over the course of the naming discussion in #10 (from selling base to selling quote tokens) and I missed to run the test plan before merging.
It turns out the current logic is incorrect and does not work for tokens with different decimals. This PR fixes this.

The orderbook API only allows us to compute the price for selling amounts denominated in the base token. The webserver API is supposed to provide prices for selling amounts denominated in the quote token.

We therefore invert the orderbook we fetch from the requests base/quote token and compute the estimated buy amount with this one. I added a comment to add the logic.

### Test Plan

http://localhost:8080/api/v1/markets/1-4/estimated-buy-amount/1000000?atoms=true

Is selling 1USDC for eth and returns the expected 6035287869762820 (~0.006 ETH).

http://localhost:8080/api/v1/markets/4-1/estimated-buy-amount/1000000000000000000?atoms=true

Is selling 1ETH for USDC and returns the expected 161143002 (~161 USDC).