Prepare release (#15)
This PR updates the underlying contracts version that fixes a bug with Fration toNumber conversion and allows us to fetch many more items per page (reducing fetch time from 40s-> 30s).

It also makes it so that we run our deployment on tags. This way I can tag this commit as v0.0.1 and issue a prod release on Tuesday (required for prod frontend).