[Easy] Use logger printing morgan request log (#44)
Currently our Kibana logs are jumbling up the log messages like this:
```
    2020-05-15 11:32:36,173 INFO [orderbook-fetcher] Updated orderbook. GET /api/v1/metrics 200 - - 0.142 ms GET /api/v1/metrics 200 - - 0.065 ms GET /api/v1/metrics 200 - - 0.063 ms GET /api/v1/metrics 200 - - 0.079 ms
```

This PR uses the same logger for the request data logs as we use internally, this will ensure that they get properly split in Kibana.

Note that we still use `morgan` instead of writing our own logging middleware (even if we make it not actually write any log messages). This is so:
- all are logs are consistent and made by the same logger
- use `morgan`'s nice uitlities for calculating request duration, content length, etc. as well as setting up all the required express hooks so things get logged correctly

### Test Plan

Run the logger and see the consistent formatting:
```
2020-05-15 13:52:10,969 INFO [dex-price-estimation] server started at http://localhost:8080
2020-05-15 13:52:19,958 INFO [dex-price-estimation] GET /api/v1/markets/2-7?atoms=1 304  - 6.113 ms
```

Make sure logs are no longer jumbled on Kibana.